From here on out, please make changes directly to the trunk and not to
the NCL_6_0_0 branch.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12092 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/AddIntrinsics.c
@@ -559,7 +559,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("integer");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
@@ -577,7 +577,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("integer");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 1;
 	args[2].n_dims = 1;
 	args[2].dim_sizes[0] = -1;
@@ -586,6 +586,7 @@ void _NclAddIntrinsics
 	args[3].dim_sizes[0] = 1;
 	args[3].n_dims = 1;
 	_NclRegisterFunc(_NclIfbindirread,args,"fbindirread",4,IFUNC);
+
 	args = NclCalloc(4,sizeof(NclArgTemplate));
 	args[0].arg_data_type = _NclLookUp("string");
 	args[0].dim_sizes[0] = 1;
@@ -595,7 +596,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("integer");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 1;
 	args[2].n_dims = 1;
 	args[2].dim_sizes[0] = -1;
@@ -627,7 +628,7 @@ void _NclAddIntrinsics
 	args[1].dim_sizes[0] = 1;
 	args[1].is_dimsizes = 1;
 	args[1].n_dims = 1;
-	args[2].arg_data_type = _NclLookUp("integer");
+	args[2].arg_data_type = _NclLookUp("numeric");
 	args[2].is_dimsizes = 0;
 	args[2].n_dims = 1;
 	args[3].arg_data_type = _NclLookUp("string");
@@ -656,7 +657,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("integer");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
@@ -670,7 +671,7 @@ void _NclAddIntrinsics
 	args[0].dim_sizes[0] = 1;
 	args[0].is_dimsizes = 1;
 	args[0].n_dims = 1;
-	args[1].arg_data_type = _NclLookUp("integer");
+	args[1].arg_data_type = _NclLookUp("numeric");
 	args[1].is_dimsizes = 0;
 	args[1].n_dims = 1;
 	args[2].arg_data_type = _NclLookUp("string");
